refactor(providers): rename types file
