+++ components/Landing/LandingQuickStart/LandingQuickStartComponent.vue
@@ -35,7 +35,7 @@
 </template>
 
 <script setup lang="ts">
-import type { ProviderObject } from "~/types/providersContent";
+import type { ProviderObject } from "~/types/providers";
 
 const { data } = await useAsyncData("providers-quick-start", () =>
   queryContent("/providers/quick-start/data").findOne()
